requirements.txt
Django==2.1.4
django-multiselectfield==0.1.8
djangorestframework==3.9.0
pytz==2018.7
